Code clean up: move function to appropriate files.
